Trace ts to f64

+++ src/proxy_common.rs
@@ -201,8 +201,8 @@ pub fn parse_bool(sbool: &str) -> bool {
 }
 
 #[allow(unused)]
-pub fn derivate_time_serie(data: &[(u64, f64)]) -> Vec<(u64, f64)> {
-    let mut ret: Vec<(u64, f64)> = vec![(data[0].0, 0.0)];
+pub fn derivate_time_serie(data: &[(f64, f64)]) -> Vec<(f64, f64)> {
+    let mut ret: Vec<(f64, f64)> = vec![(data[0].0, 0.0)];
 
     for i in (1..data.len()) {
         let deltax = (data[i].0 as f64) - (data[i - 1].0 as f64);
@@ -213,7 +213,7 @@ pub fn derivate_time_serie(data: &[(u64, f64)]) -> Vec<(u64, f64)> {
 }
 
 #[allow(unused)]
-pub fn offset_time_serie(data: &mut Vec<(u64, f64)>, offset: u64) {
+pub fn offset_time_serie(data: &mut Vec<(f64, f64)>, offset: f64) {
     for v in data {
         v.0 -= offset;
     }
